Fix several warnings on Clang with --more-warnings

+++ libcaf_core/src/ipv6_address.cpp
@@ -217,8 +217,7 @@ error parse(string_view str, ipv6_address& dest) {
   parser::read_ipv6_address(res, f);
   if (res.code == pec::success)
     return none;
-  return make_error(res.code, static_cast<size_t>(res.line),
-                    static_cast<size_t>(res.column));
+  return make_error(res.code, res.line, res.column);
 }
 
 } // namespace caf
